Flyttet Oppgave til felles (#50)
