+++ felles/src/main/kotlin/no/nav/familie/kontrakter/felles/oppgave/Oppgave.kt
@@ -1,0 +1,11 @@
+package no.nav.familie.kontrakter.felles.oppgave
+
+import javax.validation.constraints.Pattern
+
+data class Oppgave(
+    @field:Pattern(regexp = "[0-9]{13}")
+    val aktorId: String,
+    val journalpostId: String,
+    val eksisterendeOppgaveId: String?,
+    val beskrivelse: String
+)
